refactor(cognition): move from_settled test into types.rs (pure code-motion)
Relocated the self-contained SettleStep::from_settled test (and its outcome_with helper)
from mod.rs's giant test mod into types.rs's own `#[cfg(test)] mod tests` — it exercises only
the type it lives beside, needs no executor/faculty harness. Dropped the now-unused TurnMetrics
import from mod.rs's test imports. Behavior-identical. mod.rs 1597 → 1527 lines; types.rs 135 →
211 lines. act_observe tests green: 28 passed, 0 failed.
